### 2020-05-11 #43 @quenot
Reproduction of ODP-PIV from the 1998 “Particle Image Velocimetry with Optical Flow” paper
**Original article:** Quénot, Georges & Pakleza, Jaroslaw & Kowalewski, Tomasz. (1998). Particle Image Velocimetry with Optical Flow. Experiments in Fluids. Volume 25, Issue 3, pp 177–189. 10.1007/s003480050222.

**Scientific domain:**  Image Processing
**Programming language:**  ANSI C
