### 2020-05-19 #43 @rougier
Thanks for your submission. We'll assign a editor soon.

@delsuc @khinsen @pdebuyl Can any of you edit this (late) submission for the tern years reproducibilioty challenge?
