### 2020-06-15 #43 @rougier
Maybe @emmanuelle would be interested in reviewing this paper in image processing.
