### 2020-06-15 #43 @MathieuLeocmach
Hey, why not. I'm not an expert in the theory of image processing, but indeed I am chewing on exactly that problem right now. Can you point me to explanations of what is expected from a review here?
