### 2020-06-15 #43 @rougier
Here are the instruction for review https://rescience.github.io/edit/. But you can of course adapt them. Since all reviews are public, you read others at https://github.com/ReScience/submissions/issues?q=is%3Aissue to give you an idea.
