### 2020-09-09 #43 @MathieuLeocmach
Sorry, I slept on this one. I read the original paper and the paper before my holidays, and then ...

I have an issue with the way the repository is organised: the code is archived in a tgz rather than living in the git repository. Of course, I can unpack it myself, but changes or improvements will be more difficult to track. @quenot can you update your repository with the unpacked code?
