+++ jeif98/threaded/src/velo.c
@@ -1,0 +1,788 @@
+/*------------------------------------------------------------------------------------*/
+/*  Copyright (C) 1995, 1996 and 1997  Georges QUENOT  LIMSI-CNRS                     */
+/*  Copyright (C) 1998 and 1999        Georges QUENOT  CLIPS-IMAG                     */
+/*  Version 2.00 Last revision: november 16, 1999                                     */
+/*  This software comes with ABSOLUTELY NO WARRANTY                                   */
+/*                                                                                    */
+/*  Free use for educational and research purpose only of this software is possible   */
+/*  under following conditions:                                                       */
+/*  - No redistribution or inclusion in a product with or without modifications,      */
+/*  - Maintainance of the copyright notice in (modified or not) source files,         */
+/*  - Modifications and enhancements possible for local use only,                     */
+/*  - If publications are made about a system using this software or part of it,      */
+/*    the source software and author must be referenced. Citing any one paper of:     */
+/*      ftp://ftp.limsi.fr/pub/quenot/articles/icassp92.ps                            */
+/*      ftp://ftp.limsi.fr/pub/quenot/articles/imacom96.ps                            */
+/*      ftp://ftp.limsi.fr/pub/quenot/articles/mva96.ps                               */
+/*    is sufficient.                                                                  */
+/*                                                                                    */
+/*  Use within a commercial product or within the execution of a contract may         */
+/*  require a licence agreement or a cooperation convention with LIMSI-CNRS.          */
+/*------------------------------------------------------------------------------------*/
+
+#include "glob.h"
+#include "tens.h"
+#include "imio.h"
+#include "inpo.h"
+#include "velo.h"
+
+/*------------------------------------------------------------------------------------*/
+/*  intersect routine: compute (*xi,*yi) intersection of the lines:                   */
+/*  (1) passing by (x1,y1) and (x2,y2) and                                            */
+/*  (2) passing by (x,y) and directed by (vx,vy).                                     */
+/*------------------------------------------------------------------------------------*/
+
+void intersect(double x, double y, double x1, double y1, double x2, double y2,
+double vx, double vy, double *xi, double *yi)
+{
+    double xm1,ym1,x12,y12,lambda,r;
+    xm1 = x-x1;
+    ym1 = y-y1;
+    x12 = x2-x1;
+    y12 = y2-y1;
+    if (fabs(r = vx*y12-vy*x12) < 1.0e-30) r = 1.0e+30;
[…]
