+++ jeif98/threaded/src/opflow.c
@@ -1,0 +1,565 @@
+/*------------------------------------------------------------------------------------*/
+/*  Copyright (C) 1995, 1996 and 1997  Georges QUENOT  LIMSI-CNRS                     */
+/*  Copyright (C) 1998                 Georges QUENOT  CLIPS-IMAG                     */
+/*  Version 2.00 Last revision: may 25, 1998                                          */
+/*  This software comes with ABSOLUTELY NO WARRANTY                                   */
+/*                                                                                    */
+/*  Free use for educational and research purpose only of this software is possible   */
+/*  under following conditions:                                                       */
+/*  - No redistribution or inclusion in a product with or without modifications,      */
+/*  - Maintainance of the copyright notice in (modified or not) source files,         */
+/*  - Modifications and enhancements possible for local use only,                     */
+/*  - If publications are made about a system using this software or part of it,      */
+/*    the source software and author must be referenced. Citing any one paper of:     */
+/*      ftp://ftp.limsi.fr/pub/quenot/articles/icassp92.ps                            */
+/*      ftp://ftp.limsi.fr/pub/quenot/articles/imacom96.ps                            */
+/*      ftp://ftp.limsi.fr/pub/quenot/articles/mva96.ps                               */
+/*    is sufficient.                                                                  */
+/*                                                                                    */
+/*  Use within a commercial product or within the execution of a contract may         */
+/*  require a licence agreement or a cooperation convention with LIMSI-CNRS.          */
+/*------------------------------------------------------------------------------------*/
+
+#include <strings.h>
+#include "glob.h"
+#include "tens.h"
+#include "imio.h"
+#include "inpo.h"
+#include "opfl.h"
+#include "velo.h"
+
+/*------------------------------------------------------------------------------------*/
+
+#define PROGRAM "opflow"
+#define CM1     0.03
+#define CM2     1.00
+#define EPS     4.0
+#define VS      2
+
+/*------------------------------------------------------------------------------------*/
+/*--------------------- ARGUMENTS EXTRACTION AND USAGE PROCEDURES --------------------*/
+/*------------------------------------------------------------------------------------*/
+
+void usage(FILE *fp) {
+    fprintf(fp,"Usage: %s input-path output-path seq-name [-options]\n",PROGRAM);
+    fprintf(fp,"\nGeneral options:\n");
+    fprintf(fp,"  [-QUIET]              : silent mode, default is verbose\n");
+    fprintf(fp,"\nInput sequence specification parameters:\n");
+    fprintf(fp,"  [-P n]                : first frame offset relative to -M value,\n");
+    fprintf(fp,"                          default value is 1\n");
+    fprintf(fp,"  [-N n]                : last frame offset relative to -M value,\n");
[…]

[…]
