+++ jeif98/treal6.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+treal.sh original/linux
+treal.sh original/intelc
+treal.sh original/intel
+treal.sh threaded/linux
+treal.sh threaded/intelc
+treal.sh threaded/intel
