### 2020-09-10 #43 @quenot
Hello,

Thanks for the feedback. I unpacked the archives and made a few improvements, including adding the scripts for running the experiments.

Best regards,

Georges.

Le 09/09/2020 à 22:31, Mathieu Leocmach wrote:
>
> Sorry, I slept on this one. I read the original paper and the paper before my holidays, and then ...
>
> I have an issue with the way the repository is organised: the code is archived in a tgz rather than living in the git repository. Of course, I can
> unpack it myself, but changes or improvements will be more difficult to track. @quenot <https://github.com/quenot> can you update your repository
> with the unpacked code?
>
