### 2020-09-21 #43 @MathieuLeocmach
I'm still looking at it, but I suggest a few ergonomic improvements:
 - have compilation instructions in the readme.md (`make -f makelinux all` in each folder)
 - have `./` or `source ` prepended to all bash script call within bash script. Having `#!/bin/bash` should have been enough, but it's not on my platform (debian linux). For example, if I run `bat8.sh`, it's not able to call `table1-2.sh`

Edit: also, the link to the PDF file of the reproducibility paper is now broken. I think the reproducibility paper should be included in the git repository, side by side with the PDF of the original paper.
