### 2023-06-28 #43 @bkmgit
Nice article. Could run the code and obtain tables 1 and 2 from the original paper.  Can provide a further review if needed.
