### 2023-06-28 717be26
Add bash in script commands, default GNU
Ensure scripts run when called from other scripts

Assume Intel compilers may not be available by default, so
comment out and allow users to uncomment.
