+++ README.md
@@ -1,2 +1,14 @@
 # opflow
 Optical Flow using Dynamic Programming
+
+## Reproducibility using Guix
+
+- Install [Guix](https://guix.gnu.org/en/download/), either as a package
+manager on top of a linux system, or as an operating system.
+
+- Set up the compilers using
+```
+guix time-machine -C channels.scm  -m manifest.scm 
+```
+
+- Run the scripts as indicatd in the [jeif98/README.md](jeif98/README.md)
