+++ README.md
@@ -9,7 +9,7 @@ manager on top of a linux system, or as an operating system.
 - Set up the compilers using
 [Guix time-machine](https://guix.gnu.org/manual/en/html_node/Invoking-guix-time_002dmachine.html)
 ```
-guix time-machine --container channels.scm --environment --manifest manifest.scm 
+guix time-machine --container channels.scm --shell --manifest manifest.scm 
 ```
 
 - Run the scripts as indicatd in the [jeif98/README.md](jeif98/README.md)
