### 2023-07-17 #43 @quenot
Dear Benson,
This was a while ago, thanks a lot for reactivating this for me. Besides the pandemics, I had personal difficulties for some times and later I found it a bit hard to resume it. I have seen messages indicating that you worked on improving my scripts as initially requested by the reviewers, thanks also for that. I suppose that I should check and validate your modifications but am not a git expert and I am not sure about how to do it.
Best regards,
Georges.
