### 2023-07-17 #43 @bkmgit
Thanks for your message. Perhaps download the zip file at https://github.com/bkmgit/opflow/archive/refs/heads/script-updates.zip  If they seem ok, you can merge the pull request.

There is a tutorial on git at https://swcarpentry.github.io/git-novice/

For merging see https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/incorporating-changes-from-a-pull-request/merging-a-pull-request#merging-a-pull-request

The second modification uses Guix, but can be examined once bash scripts are validated to run.
